docs updated

+++ iteration/intDivision.c
@@ -16,7 +16,7 @@
  * @param b Divisor
  * @return long int Division(a/b)
  */
-long int divide(int a, int b)
+long int divide(long int a, long int b)
 {
 	unsigned long int n = 0;
 	short f = 0;
@@ -47,18 +47,18 @@ long int divide(int a, int b)
 
 int main(int argc, char *argv[])
 {	
-	int a=0, b=0;
+	long int a=0, b=0;
 	if(argc == 1){
         printf("Enter two number to calculate multiplication Ex.-3,2:\n");
-        scanf("%d,%d", &a, &b);
+        scanf("%ld,%ld", &a, &b);
     } else {
         if(!isdigit(argv[1][0]) || !isdigit(argv[2][0])) {
             printf( Yellow "Enter unsigned integer number as first & second argument "
                     "\nto calculate integer division. Example: %s" Bold_Yellow " 10 5\n", argv[0]);
             exit(1);
         }
-        a = atoi(argv[1]);
-		b = atoi(argv[2]);
+        a = atol(argv[1]);
+		b = atol(argv[2]);
     }
 	
 	printf("%ld\n",divide(a,b));
